+++ blog-vue/blog/package.json
@@ -17,7 +17,17 @@
     "highlight.js": "^10.7.3",
     "jquery": "^3.6.0",
     "markdown-it": "^11.0.1",
+    "markdown-it-abbr": "^1.0.4",
+    "markdown-it-container": "^3.0.0",
+    "markdown-it-deflist": "^2.1.0",
+    "markdown-it-emoji": "^2.0.2",
+    "markdown-it-footnote": "^3.0.3",
+    "markdown-it-ins": "^3.0.1",
+    "markdown-it-katex-external": "^1.0.0",
     "markdown-it-mark": "^3.0.1",
+    "markdown-it-sub": "^1.0.0",
+    "markdown-it-sup": "^1.0.0",
+    "markdown-it-task-lists": "^2.1.1",
     "nprogress": "^0.2.0",
     "recorderx": "^2.0.2",
     "stylus": "^0.54.7",
