+++ blog-vue/blog/public/index.html
@@ -7,9 +7,12 @@
     <!-- <meta http-equiv="Content-Security-Policy" content="upgrade-insecure-requests"> -->
     <link rel="icon" href="<%= BASE_URL %>favicon.ico">
     <title></title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/github-markdown-css/2.2.1/github-markdown.css"/>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.8.3/katex.min.css" />
     <link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900" rel="stylesheet">
     <link rel="stylesheet" href="https://fastly.jsdelivr.net/npm/@mdi/font@5.x/css/materialdesignicons.min.css" >
     <script src="https://ssl.captcha.qq.com/TCaptcha.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.8.3/katex.min.js"></script>
     <script src="http://tjs.sjs.sinajs.cn/open/api/js/wb.js?appkey=4039197195" type="text/javascript" charset="utf-8"></script>
     <script type="text/javascript" src="http://connect.qq.com/qc_jssdk.js"  data-appid="101878726" data-redirecturi="https://www.talkxj.com/oauth/login/qq" charset="utf-8"></script>
   </head>
